+++ tomcat/src/main/java/jakarta/http/HttpBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package jakarta.http;
 
 import java.util.Optional;
