+++ tomcat/src/main/java/jakarta/http/HttpRequest.java
@@ -1,8 +1,6 @@
-package org.apache.coyote.http11;
+package jakarta.http;
 
 import jakarta.servlet.http.HttpSession;
-import org.apache.catalina.Manager;
-import org.apache.catalina.session.SimpleSession;
 
 import java.io.IOException;
 import java.net.URI;
@@ -10,16 +8,17 @@ import java.net.URI;
 public class HttpRequest {
 
     private static final String REQUEST_LINE_DELIMITER = " ";
+    private static final String DEFAULT_JSSESION_ID = "";
     private static final int METHOD_POSITION = 0;
     private static final int PATH_POSITION = 1;
-    private static final String DEFAULT_JSSEION_ID = "";
 
     private final HttpMethod httpMethod;
     private final URI uri;
     private final HttpVersion httpVersion;
     private final Header header;
-    private final HttpBody body;
-    private final Manager sessionManager;
+    private final HttpBody httpBody;
+    private final HttpSessionWrapper httpSessionWrapper;
+
     private String sessionId;
 
     public HttpRequest(
@@ -27,20 +26,26 @@ public class HttpRequest {
             URI uri,
             HttpVersion httpVersion,
             Header header,
-            HttpBody body,
-            Manager sessionManager,
+            HttpBody httpBody,
+            HttpSessionWrapper httpSessionWrapper,
             String sessionId
     ) {
         this.httpMethod = httpMethod;
         this.uri = uri;
         this.httpVersion = httpVersion;
         this.header = header;
-        this.body = body;
-        this.sessionManager = sessionManager;
+        this.httpBody = httpBody;
+        this.httpSessionWrapper = httpSessionWrapper;
         this.sessionId = sessionId;
     }
 
-    public static HttpRequest createHttp11Request(String requestLine, Header header, HttpBody body, Manager manager) {
+    public static HttpRequest createHttpRequest(
+            String requestLine,
+            Header header,
+            HttpBody body,
+            HttpVersion version,
+            HttpSessionWrapper manager
+    ) {
         String[] requestLines = requestLine.split(REQUEST_LINE_DELIMITER);
         HttpMethod httpMethod = HttpMethod.from(requestLines[METHOD_POSITION]);
         URI uri = URI.create(requestLines[PATH_POSITION]);
@@ -48,11 +53,11 @@ public class HttpRequest {
         return new HttpRequest(
                 httpMethod,
                 uri,
-                HttpVersion.HTTP_1_1,
+                version,
                 header,
                 body,
                 manager,
-                header.getJSessionId().orElse(DEFAULT_JSSEION_ID)
+                header.getJSessionId().orElse(DEFAULT_JSSESION_ID)
         );
     }
 
@@ -80,19 +85,14 @@ public class HttpRequest {
         return header;
     }
 
-    public HttpBody getBody() {
-        return body;
+    public HttpBody getHttpBody() {
+        return httpBody;
     }
 
     public HttpSession getSession(boolean sessionCreateIfAbsent) throws IOException {
-        HttpSession foundSession = sessionManager.findSession(sessionId);
+        HttpSession session = httpSessionWrapper.getSession(sessionCreateIfAbsent, sessionId);
+        sessionId = session.getId();
 
-        if (foundSession == null && sessionCreateIfAbsent) {
-            HttpSession session = new SimpleSession();
-            sessionManager.add(session);
-            return session;
-        }
-
-        return foundSession;
+        return session;
     }
 }
