+++ tomcat/src/main/java/jakarta/http/HttpSessionWrapper.java
@@ -1,0 +1,10 @@
+package jakarta.http;
+
+import jakarta.servlet.http.HttpSession;
+
+import java.io.IOException;
+
+public interface HttpSessionWrapper {
+
+    HttpSession getSession(boolean sessionCreateIfAbsent, String sessionId) throws IOException;
+}
