+++ tomcat/src/main/java/jakarta/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package jakarta.http;
 
 public enum HttpStatus {
