+++ tomcat/src/test/java/jakarta/http/HeaderTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package jakarta.http;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
